+++ .gitignore
@@ -83,3 +83,7 @@ lint/generated/
 lint/outputs/
 lint/tmp/
 # lint/reports/
+/app/proguard-rules.pro
+/gradle.properties
+/settings.gradle
+/gradle/wrapper/gradle-wrapper.properties
