+++ app/src/androidTest/java/ru/igorabdulganeev/percpatyclub/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package ru.igorabdulganeev.myapplication
+package ru.igorabdulganeev.percpatyclub
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
