+++ app/src/test/java/ru/igorabdulganeev/percpatyclub/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package ru.igorabdulganeev.myapplication
+package ru.igorabdulganeev.percpatyclub
 
 import org.junit.Test
